requirements.txt
Django==3.2.15
psycopg2==2.9.3
requests==2.27.1
djangorestframework==3.13.1
